+++ CoreShell/CoreShellAssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("CoreShell.Tests")]
+[assembly: InternalsVisibleTo("CloudShell.Hosting")]