+++ CoreShell/CoreShellAttributeNames.cs
@@ -1,0 +1,10 @@
+namespace CoreShell;
+
+public static class CoreShellAttributeNames
+{
+    public const string Icon = "cloudshell.ui.composition.icon";
+
+    public const string Group = "cloudshell.ui.composition.group";
+
+    public const string GroupTitle = "cloudshell.ui.composition.groupTitle";
+}